[thrift] Patch thrift Makefile.in to link openssl in
In order to pickup the toolchain OpenSSL (and not the system version),
we patch the Makefile.in file. NOTE: I tried patching the CMakeLists.txt
but we don't seem to be regenerating the Makefiles and making that happen
surfaced more issues, so I'm going this route as it's the least invasive.
Feels hacky though.