[breakpad] Hack the configure scripts to not use getrandom
The syscall for getrandom() was added in glibc 2.25. This is available
by default on Ubuntu 18.04 but not on 16.04. If this is compiled in on
18.04, then linking for recordserviced will fail on 16.04 hosts.
To prevent this, we patch the configure script and header to not
go looking for it. This fix is likely ham-fisted, but I lack the
knowledge to do a more targeted patch and this seems functional for now.